US44882: Move tag finding function to another file for inclusion.
