+++ tests/unitTests.js
@@ -6,6 +6,8 @@ let sinon = require('sinon');
 // var proxyquire = require('proxyquire');
 let rewire = require('rewire');
 
+let common = require('../common.js');
+
 let protractorJunitXmlPlugin,
     fakeBuilder = {},
     fakeOs = {},
@@ -144,7 +146,7 @@ describe('In protractor-junit-xml-plugin', function () {
 
     describe('When XRAY-ID tag has delimiters in name field:', function () {
         it('The name is not truncated', async function () {
-            let splice = await protractorJunitXmlPlugin.findXrayIdAndName(':XRAY-ID:GLADOS-5566:Name with a colon : in it', true);
+            let splice = await common.findXrayIdAndName(':XRAY-ID:GLADOS-5566:Name with a colon : in it', true);
 
             expect(splice.xrayId).to.eq('GLADOS-5566');
             expect(splice.name).to.eq('Name with a colon : in it');
